Request not found error report : delete

+++ src/main/java/com/api/spring/boot/funsho/api/resource/usersRequestResource.java
@@ -131,6 +131,7 @@ public class usersRequestResource {
         }
         usersRequest usersRequest = UsersRequestRepository.findByRequestId(id);
 
+        if(usersRequest == null) throw new userNotFoundException("Request Not Found");
         if(user.getUserId()-usersRequest.getUserId()!=0) throw new userNotFoundException("You are not authorized to delete this request");
         
         UsersRequestRepository.deleteById(id);
